added function to check orientation of image (gives string return for simplicity)
git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/CBICA_Toolkit/trunk@569 36980a1d-918e-4cb2-ab0a-7661cd27ea28

+++ src/classes/opencv/cbicaOPENCVUtilities.cpp
@@ -91,7 +91,7 @@ namespace cbica
       while (getline(temp, single_value, ','))
       {
         // convert the string element to a integer value
-        values.push_back(atoi(single_value.c_str()));
+        values.push_back(atof(single_value.c_str()));
       }
       // add the row to the complete data vector
       if (!values.empty())
